requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-extensions==4.1

asgiref==3.11.0
sqlparse==0.5.5
tzdata==2025.3

psycopg2-binary==2.9.11

httpx==0.28.1
httpcore==1.0.9
anyio==4.12.1
h11==0.16.0
sniffio==1.3.1
idna==3.11
certifi==2026.1.4

python-dotenv==1.2.1
PyJWT==2.10.1
bcrypt==5.0.0

pydantic==2.12.5
pydantic_core==2.41.5
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspection==0.4.2

openai==2.14.0
jiter==0.12.0
tqdm==4.67.1
colorama==0.4.6
distro==1.9.0
